pyproject.toml
[project]
name = "fast-track"
version = "0.1.0"
description = "FastMCP server for expense tracking with Google Sheets integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.3",
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pydantic>=2.0.0",
]
